+++ storm-core/src/jvm/backtype/storm/scheduler/Cluster.java
@@ -90,6 +90,21 @@ public class Cluster {
         }
         this.conf = storm_conf;
     }
+
+    /**
+     * Get a copy of this cluster object
+     */
+    public static Cluster getCopy(Cluster cluster) {
+        HashMap<String, SchedulerAssignmentImpl> newAssignments = new HashMap<String, SchedulerAssignmentImpl>();
+        for (Map.Entry<String, SchedulerAssignmentImpl> entry : cluster.assignments.entrySet()) {
+            newAssignments.put(entry.getKey(), new SchedulerAssignmentImpl(entry.getValue().getTopologyId(), entry.getValue().getExecutorToSlot()));
+        }
+        Map newConf = new HashMap<String, Object>();
+        newConf.putAll(cluster.conf);
+        Cluster copy = new Cluster(cluster.inimbus, cluster.supervisors, newAssignments, newConf);
+        copy.status = new HashMap<>(cluster.status);
+        return copy;
+    }
     
     public void setBlacklistedHosts(Set<String> hosts) {
         blackListedHosts = hosts;
@@ -379,6 +394,16 @@ public class Cluster {
     }
 
     /**
+     * get slots used by a topology
+     */
+    public Collection<WorkerSlot> getUsedSlotsByTopologyId(String topologyId) {
+        if (!this.assignments.containsKey(topologyId)) {
+            return null;
+        }
+        return this.assignments.get(topologyId).getSlots();
+    }
+
+    /**
      * Get a specific supervisor with the <code>nodeId</code>
      */
     public SupervisorDetails getSupervisorById(String nodeId) {
@@ -430,10 +455,42 @@ public class Cluster {
     }
 
     /**
+     * set assignments for cluster
+     */
+    public void setAssignments(Map<String, SchedulerAssignment> newAssignments) {
+        this.assignments = new HashMap<String, SchedulerAssignmentImpl>(newAssignments.size());
+        for (Map.Entry<String, SchedulerAssignment> entry : newAssignments.entrySet()) {
+            this.assignments.put(entry.getKey(), new SchedulerAssignmentImpl(entry.getValue().getTopologyId(), entry.getValue().getExecutorToSlot()));
+        }
+    }
+
+    /**
      * Get all the supervisors.
      */
     public Map<String, SupervisorDetails> getSupervisors() {
         return this.supervisors;
+    }
+
+    /**
+     * Get the total amount of CPU resources in cluster
+     */
+    public double getClusterTotalCPUResource() {
+        double sum = 0.0;
+        for (SupervisorDetails sup : this.supervisors.values()) {
+            sum += sup.getTotalCPU();
+        }
+        return sum;
+    }
+
+    /**
+     * Get the total amount of memory resources in cluster
+     */
+    public double getClusterTotalMemoryResource() {
+        double sum = 0.0;
+        for (SupervisorDetails sup : this.supervisors.values()) {
+            sum += sup.getTotalMemory();
+        }
+        return sum;
     }
 
     /*
@@ -560,12 +617,26 @@ public class Cluster {
         }
     }
 
+    /**
+     * set scheduler status for a topology
+     */
     public void setStatus(String topologyId, String status) {
         this.status.put(topologyId, status);
     }
 
+    /**
+     * Get all topology scheduler statuses
+     */
     public Map<String, String> getStatusMap() {
         return this.status;
+    }
+
+    /**
+     * set scheduler status map
+     */
+    public void setStatusMap(Map<String, String> statusMap) {
+        this.status.clear();
+        this.status.putAll(statusMap);
     }
 
     public void setResources(String topologyId, Double[] resources) {