+++ storm-core/src/jvm/backtype/storm/scheduler/resource/strategies/scheduling/DefaultResourceAwareStrategy.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package backtype.storm.scheduler.resource.strategies;
+package backtype.storm.scheduler.resource.strategies.scheduling;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -30,6 +30,10 @@ import java.util.TreeMap;
 import java.util.HashSet;
 import java.util.Iterator;
 
+import backtype.storm.scheduler.resource.RAS_Nodes;
+import backtype.storm.scheduler.resource.SchedulingResult;
+import backtype.storm.scheduler.resource.SchedulingStatus;
+import backtype.storm.scheduler.resource.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -41,8 +45,8 @@ import backtype.storm.scheduler.WorkerSlot;
 import backtype.storm.scheduler.resource.Component;
 import backtype.storm.scheduler.resource.RAS_Node;
 
-public class ResourceAwareStrategy implements IStrategy {
-    private static final Logger LOG = LoggerFactory.getLogger(ResourceAwareStrategy.class);
+public class DefaultResourceAwareStrategy implements IStrategy {
+    private static final Logger LOG = LoggerFactory.getLogger(DefaultResourceAwareStrategy.class);
     private Topologies _topologies;
     private Cluster _cluster;
     //Map key is the supervisor id and the value is the corresponding RAS_Node Object 
@@ -58,10 +62,10 @@ public class ResourceAwareStrategy implements IStrategy {
     private final double MEM_WEIGHT = 1.0;
     private final double NETWORK_WEIGHT = 1.0;
 
-    public ResourceAwareStrategy(Cluster cluster, Topologies topologies) {
+    public void prepare (Topologies topologies, Cluster cluster, Map<String, User> userMap, RAS_Nodes nodes) {
         _topologies = topologies;
         _cluster = cluster;
-        _nodes = RAS_Node.getAllNodesFrom(cluster, _topologies);
+        _nodes = RAS_Nodes.getAllNodesFrom(cluster, _topologies);
         _availNodes = this.getAvailNodes();
         _clusterInfo = cluster.getNetworkTopography();
         LOG.debug(this.getClusterInfo());
@@ -84,10 +88,10 @@ public class ResourceAwareStrategy implements IStrategy {
         return retMap;
     }
 
-    public Map<WorkerSlot, Collection<ExecutorDetails>> schedule(TopologyDetails td) {
+    public SchedulingResult schedule(TopologyDetails td) {
         if (_availNodes.size() <= 0) {
             LOG.warn("No available nodes to schedule tasks on!");
-            return null;
+            return SchedulingResult.failure(SchedulingStatus.FAIL_NOT_ENOUGH_RESOURCES, "No available nodes to schedule tasks on!");
         }
         Collection<ExecutorDetails> unassignedExecutors = _cluster.getUnassignedExecutors(td);
         Map<WorkerSlot, Collection<ExecutorDetails>> schedulerAssignmentMap = new HashMap<>();
@@ -97,7 +101,7 @@ public class ResourceAwareStrategy implements IStrategy {
 
         if (spouts.size() == 0) {
             LOG.error("Cannot find a Spout!");
-            return null;
+            return SchedulingResult.failure(SchedulingStatus.FAIL_INVALID_TOPOLOGY, "Cannot find a Spout!");
         }
 
         Queue<Component> ordered__Component_list = bfs(_topologies, td, spouts);
@@ -159,18 +163,23 @@ public class ResourceAwareStrategy implements IStrategy {
                 LOG.error("Not Enough Resources to schedule Task {}", exec);
             }
         }
+
+        SchedulingResult result;
         executorsNotScheduled.removeAll(scheduledTasks);
         if (executorsNotScheduled.size() > 0) {
             LOG.error("Not all executors successfully scheduled: {}",
                     executorsNotScheduled);
             schedulerAssignmentMap = null;
+            result = SchedulingResult.failure(SchedulingStatus.FAIL_NOT_ENOUGH_RESOURCES,
+                    (td.getExecutors().size() - unassignedExecutors.size()) + "/" + td.getExecutors().size() + " executors scheduled");
         } else {
             LOG.debug("All resources successfully scheduled!");
+            result = SchedulingResult.successWithMsg(schedulerAssignmentMap, "Fully Scheduled by DefaultResourceAwareStrategy");
         }
         if (schedulerAssignmentMap == null) {
             LOG.error("Topology {} not successfully scheduled!", td.getId());
         }
-        return schedulerAssignmentMap;
+        return result;
     }
 
     private WorkerSlot findWorkerForExec(ExecutorDetails exec, TopologyDetails td, Map<WorkerSlot, Collection<ExecutorDetails>> scheduleAssignmentMap) {